[修改] 修改 ins_user.USER_ID|CREATE_USER_ID|UPDATE_USER_ID 类型

+++ src/main/java/com/microtomato/hirun/modules/user/controller/UserController.java
@@ -58,7 +58,7 @@ public class UserController {
         }
 
         UserContext userContext = WebContextUtils.getUserContext();
-        Integer userId = userContext.getUserId();
+        Long userId = userContext.getUserId();
 
         boolean changeResult = userServiceImpl.changeStaffPassword(userId, oldPassword, repassword);
